+++ src/collective/cover/tests/test_vocabularies.py
@@ -43,9 +43,6 @@ class VocabulariesTestCase(unittest.TestCase):
         self.assertIn(u'collective.cover.file', tiles)
         self.assertIn(u'collective.cover.list', tiles)
         self.assertIn(u'collective.cover.richtext', tiles)
-        # XXX: tiles are now deprecated
-        self.assertNotIn(u'collective.cover.image', tiles)
-        self.assertNotIn(u'collective.cover.link', tiles)
 
     def test_enabled_tiles_vocabulary(self):
         name = u'collective.cover.EnabledTiles'
@@ -64,9 +61,6 @@ class VocabulariesTestCase(unittest.TestCase):
         self.assertIn(u'collective.cover.richtext', tiles)
         # FIXME see: https://github.com/collective/collective.cover/issues/194
         self.assertIn(u'collective.cover.pfg', tiles)
-        # XXX: tiles are now deprecated
-        self.assertNotIn(u'collective.cover.image', tiles)
-        self.assertNotIn(u'collective.cover.link', tiles)
         # XXX: https://github.com/collective/collective.cover/issues/81
         # standard tiles are not enabled... yet
         self.assertNotIn(u'plone.app.imagetile', tiles)